+++ lib/builtin/search/search.dart
@@ -1,0 +1,3 @@
+export 'src/extension.dart';
+export 'src/find_in_files_controller.dart';
+export 'src/search_panel_view.dart';